requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
scikit-learn>=1.0.0
Pillow>=8.0.0
streamlit>=1.10.0
flask>=2.0.0 